Add files via upload
Whatup guys . Was talking to Thang and we thought wed do a github account where we all post our code for this last project. I attached my code for the HW 23/19 edit. I still have to work on it cuz it sucks rn butfigured over the weekend i can work it out. Gonna share this link with all of you.